requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
black
isort
python-dotenv
